+++ blkshp_os/recipes/doctype/allergen/__init__.py
@@ -1,0 +1,5 @@
+"""
+Allergen DocType package.
+"""
+
+