+++ blkshp_os/recipes/doctype/recipe_allergen/__init__.py
@@ -1,0 +1,5 @@
+"""
+Recipe Allergen child table package.
+"""
+
+