+++ blkshp_os/recipes/doctype/recipe_inherited_allergen/__init__.py
@@ -1,0 +1,5 @@
+"""
+Recipe Inherited Allergen child table package.
+"""
+
+